+++ kmock-processor/src/test/resources/mock/expected/overloaded/Refined.kt
@@ -123,7 +123,7 @@ internal class RefinedMock(
         ProxyFactory.createSyncFunProxy("mock.template.overloaded.RefinedMock#_fooWithAlias",
             collector = collector, freeze = freeze)
 
-    public val _fooWithRList_T: KMockContract.SyncFunProxy<Unit, (List<Any?>) -> Unit> =
+    public val _fooWithRList_T: KMockContract.SyncFunProxy<Unit, (List<Any>) -> Unit> =
         ProxyFactory.createSyncFunProxy("mock.template.overloaded.RefinedMock#_fooWithRList_T",
             collector = collector, freeze = freeze)