+++ kmock-processor/src/test/resources/mock/expected/typealiaz/Platform.kt
@@ -3,7 +3,6 @@ package mock.template.typealiaz
 import kotlin.Any
 import kotlin.Boolean
 import kotlin.Comparable
-import kotlin.Function1
 import kotlin.Suppress
 import kotlin.Unit
 import tech.antibytes.kmock.KMockContract
@@ -37,18 +36,17 @@ internal class PlatformMock<L : Alias23>(
         ProxyFactory.createSyncFunProxy("mock.template.typealiaz.PlatformMock#_doAnythingElseWithAlias77",
             collector = collector, freeze = freeze)
 
-    public val _doOtherThing:
-        KMockContract.SyncFunProxy<Unit, (Alias77<Generic<Function1<Any, Unit>>>,
-            Comparable<Any?>) -> Unit> =
+    public val _doOtherThing: KMockContract.SyncFunProxy<Unit, (Alias77<Alias77<Alias21>>,
+        Comparable<Any>) -> Unit> =
         ProxyFactory.createSyncFunProxy("mock.template.typealiaz.PlatformMock#_doOtherThing",
             collector = collector, freeze = freeze)
 
-    public val _doSomethingElseWithTAlias21LAlias23:
-        KMockContract.SyncFunProxy<Unit, (Function1<Any, Unit>, Function1<Any, Any>) -> Unit> =
+    public val _doSomethingElseWithTAlias21LAlias23: KMockContract.SyncFunProxy<Unit, (Alias21,
+        Alias23) -> Unit> =
         ProxyFactory.createSyncFunProxy("mock.template.typealiaz.PlatformMock#_doSomethingElseWithTAlias21LAlias23",
             collector = collector, freeze = freeze)
 
-    public val _doSomethingElseWithTAlias77: KMockContract.SyncFunProxy<Unit, (Generic<Any?>) -> Unit>
+    public val _doSomethingElseWithTAlias77: KMockContract.SyncFunProxy<Unit, (Alias77<Any?>) -> Unit>
         =
         ProxyFactory.createSyncFunProxy("mock.template.typealiaz.PlatformMock#_doSomethingElseWithTAlias77",
             collector = collector, freeze = freeze)