+++ kmock-processor/src/test/resources/mock/template/generic/Shared.kt
@@ -86,4 +86,6 @@ interface Shared<K, L> where L : Any, L : Comparable<L>, K : Any {
 
     fun <R, T> rrr(arg0: T, arg1: R) where R : Sequence<T>, T : List<R>
     fun <R, T> rol(arg0: T, arg1: R) where R : Sequence<T>, T : List<Array<out Int>>
+    fun <R, T> lol(arg0: T, arg1: R) where R : Sequence<T>, T : List<Array<in Int>>
+    fun <R, T> nol(arg0: T, arg1: R) where R : Sequence<T>, T : List<Array<out Any?>>
 }