+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * vararg eats specialised Array Types (e.g. IntArray) and covariant types when inherited
 * Factories for Multi-Interface-Mocks
+* Multi-Boundary Parameter are not right resolved when mixed multi with regular parameter in nested types
 
 ### Security